game starts when start game btn is clicked by host
