+++ src/pages/HostGame.js
@@ -1,6 +1,6 @@
 
 import React, { useEffect, useState } from "react";
-import Lobby from "../pages/Lobby";
+import Lobby from "../components/Lobby";
 import { v4 as uuidv4 } from 'uuid';
 import "../styles/NewGame.css";
